+++ src/Instrumentum_sanae_doctrinae/web_scraping/monergism/scripture/mn_scrap_scripture_works.py
@@ -9,90 +9,15 @@ from bs4 import BeautifulSoup
 from Instrumentum_sanae_doctrinae.web_scraping import http_connexion, my_constants
 from Instrumentum_sanae_doctrinae.web_scraping.monergism import mn_scrap_metadata
 from Instrumentum_sanae_doctrinae.my_tools import general_tools as _my_tools
-from Instrumentum_sanae_doctrinae.web_scraping.monergism.mn_scrap_subtopic_work import MN_ScriptureSubtopicWork,get_subtopics, scrap_page_works
+from Instrumentum_sanae_doctrinae.web_scraping.monergism.mn_scrap_subtopic_work import MN_ScrapScriptureOrTopicWork
 
 
 
-class MN_ScriptureWork(mn_scrap_metadata.MonergismScrapAuthorTopicScripturePage):
-    def __init__(self, name, root_folder, url_list, browse_by_type,intermdiate_folders = None):
-
-        super().__init__(name, root_folder, url_list, browse_by_type,
-                          information_type_root_folder = my_constants.WORK_INFORMATION_ROOT_FOLDER,
-                          intermdiate_folders = intermdiate_folders)
+class MN_ScriptureWork(MN_ScrapScriptureOrTopicWork):
+    def  __init__(self, name, root_folder, url_list, browse_by_type, intermdiate_folders=None):
+        super().__init__(name, root_folder, url_list, browse_by_type, intermdiate_folders)
         
-
     
-    
-    async def scrap_url_pages(self):
-        """
-        Scrap the main links of the page. See this file (documentation/documentation.odt) for more info 
-        """
-
-        final_result = {}    
-        
-
-        for main_url in self.url_informations:
-            
-            # Take the div containing the links of the author 
-            
-            bs4_object = self.url_informations[main_url].get("bs4_object")
-            
-            
-            sub_url_topics_list = get_subtopics(bs4_object,main_url)
-            
-            
-            for url_info in sub_url_topics_list:
-                #print(url_info,"\n\n\n")
-                intermediate_folders = self.intermdiate_folders[
-                self.intermdiate_folders.index(my_constants.WORK_INFORMATION_ROOT_FOLDER) + 1:]
-                
-                #print(self.intermdiate_folders,"primary",intermediate_folders)
-                
-                                
-                ob = MN_ScriptureSubtopicWork(name = self.name,root_folder=self.root_folder,
-                                              url_list=[{"url":url_info.get("url")}],browse_by_type=self.browse_by_type,
-                                              intermdiate_folders= 
-                                              intermediate_folders + ["subtopics",url_info.get("name")])
-                
-                await ob.scrap_and_write(save_html_file=True)
-            
-            final_result[main_url] = scrap_page_works(bs4_object)
-
-        return final_result
-    
-    async def is_data_downloaded(self):
-
-        for url in self.url_informations:
-            file_path = self.url_informations[url].get("json_filepath")
-            
-            if not os.path.exists(file_path):
-                return False
-                        
-            file_content = await _my_tools.async_read_file(file_path)
-            
-            if not file_content:
-                return False 
-            
-            try:
-                file_content = json.loads(file_content)
-            except:
-                return False
-            
-            
-            if not file_content:
-                return False
-            
-            # Check mandatory information in the json file 
-            if not file_content.get("url"):
-                return False
-            
-
-            if not file_content.get("data"):
-                return False
-            
-        return True
-
-
 
 
 class MN_ScriptureWork_All(http_connexion.ParallelHttpConnexionWithLogManagement):
@@ -234,9 +159,8 @@ class MN_ScriptureWork_All(http_connexion.ParallelHttpConnexionWithLogManagement
                 url_list = [{"url":i} for i in element.get("data").get("pages")],
                 intermdiate_folders = element.get("download_log").get("intermediate_folders")
             )
-            if not ob.is_data_downloaded():
+            if not await ob.is_data_downloaded():
                 return False 
         
         return True 
         
-
